requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
linkedin-jobs-scraper==2.0.7
MarkupSafe==2.1.2
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
polars==0.16.14
pyarrow==11.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
selenium==3.141.0
six==1.16.0
tabulate==0.9.0
tqdm==4.65.0
urllib3==1.26.15
webdriver-manager==3.8.5
websocket-client==0.59.0
Werkzeug==2.2.3
xlsx2csv==0.8.1
XlsxWriter==3.0.9
python-decouple
